+++ src/main/java/io/github/techtastic/oc2rcreate/device/block/AbstractBlockRPCDevice.java
@@ -3,6 +3,7 @@ package io.github.techtastic.oc2rcreate.device.block;
 import li.cil.oc2.api.bus.device.object.ObjectDevice;
 import li.cil.oc2.api.bus.device.rpc.RPCDevice;
 import li.cil.oc2.api.bus.device.rpc.RPCMethodGroup;
+import org.jetbrains.annotations.NotNull;
 
 import java.util.List;
 
@@ -13,11 +14,11 @@ public class AbstractBlockRPCDevice implements RPCDevice {
         this.device = new ObjectDevice(this, typeName);
     }
 
-    public List<String> getTypeNames() {
+    public @NotNull List<String> getTypeNames() {
         return this.device.getTypeNames();
     }
 
-    public List<RPCMethodGroup> getMethodGroups() {
+    public @NotNull List<RPCMethodGroup> getMethodGroups() {
         return this.device.getMethodGroups();
     }
 }